--- a/LightResults.sln
+++ b/LightResults.sln
@@ -4,6 +4,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LightResults", "src\LightRe
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LightResults.Tests", "tests\LightResults.Tests\LightResults.Tests.csproj", "{93C8BC67-7F34-4935-A671-F0B12AEDB072}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LightResults.CurrentBenchmarks", "tests\LightResults.CurrentBenchmarks\LightResults.CurrentBenchmarks.csproj", "{EC891C69-F904-43B6-83DF-F873F5189A42}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LightResults.DevelopBenchmarks", "tests\LightResults.DevelopBenchmarks\LightResults.DevelopBenchmarks.csproj", "{E96CA1A7-5E74-4641-A62B-3C05FFCC3E4F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +22,13 @@ Global
 		{93C8BC67-7F34-4935-A671-F0B12AEDB072}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93C8BC67-7F34-4935-A671-F0B12AEDB072}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93C8BC67-7F34-4935-A671-F0B12AEDB072}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC891C69-F904-43B6-83DF-F873F5189A42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC891C69-F904-43B6-83DF-F873F5189A42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC891C69-F904-43B6-83DF-F873F5189A42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC891C69-F904-43B6-83DF-F873F5189A42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E96CA1A7-5E74-4641-A62B-3C05FFCC3E4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E96CA1A7-5E74-4641-A62B-3C05FFCC3E4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E96CA1A7-5E74-4641-A62B-3C05FFCC3E4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E96CA1A7-5E74-4641-A62B-3C05FFCC3E4F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal