--- a/tests/LightResults.CurrentBenchmarks/Program.cs
+++ b/tests/LightResults.CurrentBenchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using LightResults.CurrentBenchmarks;
+
+BenchmarkRunner.Run<Benchmarks>();