--- a/tests/LightResults.DevelopBenchmarks/Program.cs
+++ b/tests/LightResults.DevelopBenchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using LightResults.DevelopBenchmarks;
+
+BenchmarkRunner.Run<Benchmarks>();