```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3155/23H2/2023Update/SunValley3)
13th Gen Intel Core i7-13700KF, 1 CPU, 24 logical and 16 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX2
  Iterations : 100 000
```
| Method                        |     Before |     After | Difference |   Before |   After | Difference |
|-------------------------------|-----------:|----------:|------------|---------:|--------:|------------|
| ResultBase.HasError<IError>() | 4,839.8 μs |  75.49 μs | 98% faster | 12.21 MB |       - | Eliminated |
| Result.Ok()                   |   168.4 μs |  18.79 μs | 89% faster |  2.29 MB |       - | Eliminated |
| Result.Fail()                 | 1,195.6 μs |  18.98 μs | 98% faster | 12.21 MB |       - | Eliminated |
| Result.Fail(IError)           | 1,029.6 μs | 363.17 μs | 65% faster |  9.16 MB | 5.34 MB | 42% less   |
| Result.Ok<TValue>(TValue)     |   195.3 μs | 310.73 μs | 59% slower |  3.05 MB | 3.05 MB | No change  |
| Result.Fail<TValue>()         | 1,251.3 μs |  18.94 μs | 98% faster | 12.97 MB |       - | Eliminated |
| Result.Fail<TValue>(IError)   | 1,013.1 μs | 382.84 μs | 62% faster |  9.92 MB | 6.10 MB | 39% less   |
| Result<TValue>.Ok(TValue)     |   196.3 μs | 305.62 μs | 56% slower |  3.05 MB | 3.05 MB | No change  |
| Result<TValue>.Fail()         | 1,225.4 μs |  18.91 μs | 98% faster | 12.97 MB |       - | Eliminated |
| Result<TValue>.Fail(IError)   | 1,018.3 μs | 380.92 μs | 63% faster |  9.92 MB | 6.10 MB | 39% less   |
| new Error()                   |   225.8 μs | 225.99 μs | No change  |  3.05 MB | 3.05 MB | No change  |
| new Error(errorMessage)       |   224.4 μs | 228.76 μs | No change  |  3.05 MB | 3.05 MB | No change  |